+++ hyoga/plot/naturalearth.py
@@ -17,7 +17,7 @@ import hyoga.plot
 # Natural Earth internals
 # -----------------------
 
-def add_feature(category=None, name=None, scale='10m', **kwargs):
+def feature(category=None, name=None, scale='10m', **kwargs):
     """Plot Natural Earth feature allowing a different color for the
     subject."""
     fname = cshp.natural_earth(resolution=scale, category=category, name=name)
@@ -27,8 +27,8 @@ def add_feature(category=None, name=None, scale='10m', **kwargs):
 # Natural Earth cultural
 # ----------------------
 
-def add_cities(ax=None, lang=None, include=None, exclude=None, ranks=None,
-               **kwargs):
+def cities(ax=None, lang=None, include=None, exclude=None, ranks=None,
+           **kwargs):
     """
     Plot populated places as an annotated scatter plot.
 
@@ -96,38 +96,38 @@ def add_cities(ax=None, lang=None, include=None, exclude=None, ranks=None,
         transform=crs, **kwargs)
 
 
-def add_countries(edgecolor='none', facecolor='#e0e0e0', linewidth=1.0,
-                  subject=None, subject_facecolor='#fefee9', **kwargs):
-    return add_feature(
+def countries(edgecolor='none', facecolor='#e0e0e0', linewidth=1.0,
+              subject=None, subject_facecolor='#fefee9', **kwargs):
+    return feature(
         category='cultural', name='admin_0_countries',
         edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
         subject=subject, subject_facecolor=subject_facecolor, **kwargs)
 
 
-def add_country_borders(edgecolor='#646464', facecolor='none', linewidth=2.0,
-                        **kwargs):
+def country_borders(edgecolor='#646464', facecolor='none', linewidth=2.0,
+                    **kwargs):
     return (
-        add_feature(
+        feature(
             category='cultural', name='admin_0_boundary_lines_land',
             edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
             **kwargs),
-        add_feature(
+        feature(
             category='cultural', name='admin_0_boundary_lines_map_units',
             edgecolor=edgecolor, facecolor=facecolor, linewidth=0.75*linewidth,
             **kwargs))
 
 
-def add_states(edgecolor='none', facecolor='#e0e0e0', linewidth=0.25,
-               subject=None, subject_facecolor='#fefee9', **kwargs):
-    return add_feature(
+def states(edgecolor='none', facecolor='#e0e0e0', linewidth=0.25,
+           subject=None, subject_facecolor='#fefee9', **kwargs):
+    return feature(
         category='cultural', name='admin_1_states_provinces',
         edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
         subject=subject, subject_facecolor=subject_facecolor, **kwargs)
 
 
-def add_state_borders(edgecolor='#646464', facecolor='none', linewidth=1.0,
-                      **kwargs):
-    return add_feature(
+def state_borders(edgecolor='#646464', facecolor='none', linewidth=1,
+                  **kwargs):
+    return feature(
         category='cultural', name='admin_1_states_provinces_lines',
         edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
         **kwargs)
@@ -136,57 +136,53 @@ def add_state_borders(edgecolor='#646464', facecolor='none', linewidth=1.0,
 # Natural Earth physical
 # ----------------------
 
-def add_coastline(edgecolor='#0978ab', facecolor='none', linewidth=0.25,
-                  **kwargs):
-    return add_feature(
+def coastline(edgecolor='#0978ab', facecolor='none', linewidth=0.25, **kwargs):
+    return feature(
         category='physical', name='coastline',
         edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
         **kwargs)
 
 
-def add_glaciers(edgecolor='#0978ab', facecolor='#ffffff', linewidth=0.25,
-                 **kwargs):
-    return add_feature(
+def glaciers(edgecolor='#0978ab', facecolor='#ffffff', linewidth=0.25,
+             **kwargs):
+    return feature(
         category='physical', name='glaciated_areas',
         edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
         **kwargs)
 
 
-def add_lakes(edgecolor='#0978ab', facecolor='#d8f2fe', linewidth=0.25,
-              **kwargs):
+def lakes(edgecolor='#0978ab', facecolor='#d8f2fe', linewidth=0.25, **kwargs):
     kwargs = dict(category='physical', edgecolor=edgecolor,
                   facecolor=facecolor, linewidth=linewidth, **kwargs)
-    features = add_feature(name='lakes', **kwargs)
+    features = feature(name='lakes', **kwargs)
     if 'scale' not in kwargs or kwargs['scale'] == '10m':
         features = (features,
-                    add_feature(name='lakes_europe', **kwargs),
-                    add_feature(name='lakes_north_america', **kwargs))
+                    feature(name='lakes_europe', **kwargs),
+                    feature(name='lakes_north_america', **kwargs))
     return features
 
 
-def add_ocean(edgecolor='#0978ab', facecolor='#c6ecff', linewidth=0.25,
-              **kwargs):
-    return add_feature(
+def ocean(edgecolor='#0978ab', facecolor='#c6ecff', linewidth=0.25, **kwargs):
+    return feature(
         category='physical', name='ocean',
         edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
         **kwargs)
 
 
-def add_rivers(edgecolor='#0978ab', facecolor='none', linewidth=0.5,
-               **kwargs):
+def rivers(edgecolor='#0978ab', facecolor='none', linewidth=0.5, **kwargs):
     kwargs = dict(category='physical', edgecolor=edgecolor,
                   facecolor=facecolor, linewidth=linewidth, **kwargs)
-    features = add_feature(name='rivers_lake_centerlines', **kwargs)
+    features = feature(name='rivers_lake_centerlines', **kwargs)
     if 'scale' not in kwargs or kwargs['scale'] == '10m':
         features = (features,
-                    add_feature(name='rivers_europe', **kwargs),
-                    add_feature(name='rivers_north_america', **kwargs))
+                    feature(name='rivers_europe', **kwargs),
+                    feature(name='rivers_north_america', **kwargs))
     return features
 
 
-def add_graticules(edgecolor='0.25', facecolor='none', linewidth=0.1,
-                   interval=1, **kwargs):
-    return add_feature(
+def graticules(edgecolor='0.25', facecolor='none', linewidth=0.1, interval=1,
+               **kwargs):
+    return feature(
         category='physical', name='graticules_{}'.format(interval),
         edgecolor=edgecolor, facecolor=facecolor, linewidth=linewidth,
         **kwargs)